requirements.txt

Django>=5.0,<6
djangorestframework>=3.15
drf-spectacular>=0.27
drf-spectacular-sidecar>=2024.7.1
